--- a/index.js
+++ b/index.js
@@ -2,13 +2,7 @@ var binding = require("./lib/jsdawg.node");
 var assert = require("assert");
 var Symbol = require("es6-symbol");
 
-binding.Dawg.prototype.toCompactDawg = function() {
-    return new CompactDawg(this.toCompactDawgBuffer());
-}
-
-var CompactDawg = function(buf) {
-    this.data = buf;
-
+function validate(buf) {
     // validate data
     var magic = buf.slice(0, 4).toString();
     var version = buf[4];
@@ -29,17 +23,30 @@ var CompactDawg = function(buf) {
     assert(offsetWidth == 4, "only dawgs with four-byte offset widths are supported");
     assert(size == actualSize, "dawg size is not as expected");
     assert(checksum == actualChecksum, "dawg checksum is not correct");
+    return buf;
 }
 
-CompactDawg.prototype.lookupPrefix = function(prefix) {
-    return binding.compactDawgBufferLookup(this.data, prefix) != 0;
+binding.Dawg.prototype.toCompactDawg = function() {
+    return new binding.CompactDawg(validate(this.toCompactDawgBuffer()));
 }
 
-CompactDawg.prototype.lookup = function(prefix) {
-    return binding.compactDawgBufferLookup(this.data, prefix) == 2;
+binding.Dawg.prototype.toCompactDawgIterator = function() {
+    return new CompactDawgIterator(validate(this.toCompactDawgBuffer()));
 }
 
-CompactDawg.prototype.iterator = function(prefix) {
+binding.CompactDawg.prototype.lookupPrefix = function(prefix) {
+    return this._lookup(prefix) != 0;
+}
+
+binding.CompactDawg.prototype.lookup = function(prefix) {
+    return this._lookup(prefix) == 2;
+}
+
+var CompactDawgIterator = function(buf) {
+    this.data = buf;
+}
+
+CompactDawgIterator.prototype.iterator = function(prefix) {
     // implement the ES6 iterator pattern
     var it = prefix ? new binding.CompactDawgIterator(this.data, prefix) : new binding.CompactDawgIterator(this.data);
     return {
@@ -57,9 +64,10 @@ CompactDawg.prototype.iterator = function(prefix) {
     }
 }
 
-CompactDawg.prototype[Symbol.iterator] = CompactDawg.prototype.iterator;
+CompactDawgIterator.prototype[Symbol.iterator] = CompactDawgIterator.prototype.iterator;
 
 module.exports = {
     Dawg: binding.Dawg,
-    CompactDawg: CompactDawg
+    CompactDawg: binding.CompactDawg,
+    CompactDawgIterator: CompactDawgIterator
 };