--- a/src/binding.cpp
+++ b/src/binding.cpp
@@ -60,6 +60,9 @@ class JSDawg : public Nan::ObjectWrap {
     }
 
     static NAN_METHOD(Insert) {
+        if (!info[0]->IsString()) {
+            return Nan::ThrowTypeError("first argument must be a String");
+        }
         JSDawg* obj = Nan::ObjectWrap::Unwrap<JSDawg>(info.This());
         String::Utf8Value utf8_value(info[0].As<String>());
         std::string input = std::string(*utf8_value, utf8_value.length());
@@ -76,6 +79,9 @@ class JSDawg : public Nan::ObjectWrap {
     }
 
     static NAN_METHOD(Lookup) {
+        if (!info[0]->IsString()) {
+            return Nan::ThrowTypeError("first argument must be a String");
+        }
         JSDawg* obj = Nan::ObjectWrap::Unwrap<JSDawg>(info.This());
         String::Utf8Value utf8_value(info[0].As<String>());
         std::string input = std::string(*utf8_value, utf8_value.length());
@@ -85,6 +91,9 @@ class JSDawg : public Nan::ObjectWrap {
     }
 
     static NAN_METHOD(LookupPrefix) {
+        if (!info[0]->IsString()) {
+            return Nan::ThrowTypeError("first argument must be a String");
+        }
         JSDawg* obj = Nan::ObjectWrap::Unwrap<JSDawg>(info.This());
         String::Utf8Value utf8_value(info[0].As<String>());
         std::string input = std::string(*utf8_value, utf8_value.length());
@@ -188,7 +197,19 @@ dawg_search_result compact_dawg_search(unsigned char* data, unsigned char* searc
 }
 
 NAN_METHOD(CompactLookup) {
+    if (!info[0]->IsObject()) {
+        return Nan::ThrowTypeError("first argument must be a Buffer");
+    }
+
     v8::Local<v8::Object> bufferObj = info[0]->ToObject();
+
+    if (bufferObj->IsNull() || bufferObj->IsUndefined() || !node::Buffer::HasInstance(bufferObj)) {
+        return Nan::ThrowTypeError("first argument must be a Buffer");
+    }
+
+    if (!info[1]->IsString()) {
+        return Nan::ThrowTypeError("second argument must be a String");
+    }
     String::Utf8Value utf8_value(info[1].As<String>());
 
     unsigned char* search = (unsigned char*) *utf8_value;
@@ -398,16 +419,8 @@ NAN_METHOD(Crc32c) {
 static NAN_MODULE_INIT(Init) {
     JSDawg::Init(target);
     CompactIterator::Init(target);
-    Nan::Set(
-        target,
-        Nan::New("compactDawgBufferLookup").ToLocalChecked(),
-        Nan::GetFunction(New<FunctionTemplate>(CompactLookup)).ToLocalChecked()
-    );
-    Nan::Set(
-        target,
-        Nan::New("crc32c").ToLocalChecked(),
-        Nan::GetFunction(New<FunctionTemplate>(Crc32c)).ToLocalChecked()
-    );
+    Nan::SetMethod(target,"compactDawgBufferLookup",CompactLookup);
+    Nan::SetMethod(target,"crc32c",Crc32c);
 }
 
 NODE_MODULE(jsdawg, Init)