Perf improvements:

 - Avoids or reduces `std::string` allocation in several places
 - Reduces unnecessary reference counting of `std::shared_ptr` by avoiding unneeded copies of `DawgNodeCheckEntry`
 - Uses `std::move` to move strings instead of copying

Other non-performance focused improvements:

 - Use C++11 `for` loops for better readability
 - Starts validating arguments