@springmeyer 2017-01-26
This PR is ready. However, after discussing with @apendleton today about how the `compactDawg.lookupPrefix` and `compactDawg.lookup` functions are the most performance critical... I have ideas on speeding those up too.

Profiling of `node bench/compact-bench.js` shows that the conversion of `v8::String` and `Node::Buffer` into C++ are major bottlenecks. It looks possible to optimize these conversions further:

1) Avoid repeated calls to `node::Buffer::Data`

Currently `node::Buffer::Data` is called in C++ every time the JS `compactDawg.lookupPrefix` or `compactDawg.lookup` functions are called. This is because those functions wrap the standalone C++ `CompactLookup` function and pass a node `Buffer` from JS into C++ every call. Calling `node::Buffer::Data` is pretty cheap in general but still shows up as a significant bottleneck:
 
![screen shot 2017-01-25 at 6 40 55 pm](https://cloud.githubusercontent.com/assets/20300/22318248/ecc1f972-e32d-11e6-9b39-5cc56f36d9f9.png)

My plan is to create a new C++ object called `CompactDawg` that will be created from a JS `Buffer` and will store a `v8::Persistant` to that buffer. Then the lookup calls can use C++ memory via a member rather than needing to pass the JS Buffer across the JS -> C++ boundary every time these wrapper functions are called: https://github.com/mapbox/dawg-cache/blob/master/index.js#L34-L40.

2) Optimize v8::String -> `const char *` conversion

Currently we are using `v8::String::Utf8Value` to convert JS strings to C++ memory.

  - Definition: https://github.com/v8/v8/blob/c6f925d7ed002e4bbd57a674cf90627384c81964/include/v8.h#L2639
  - Implementation: https://github.com/v8/v8/blob/c6f925d7ed002e4bbd57a674cf90627384c81964/src/api.cc#L8771

Profiling shows this is a major bottleneck:

![screen shot 2017-01-25 at 6 33 17 pm](https://cloud.githubusercontent.com/assets/20300/22318218/be5afcb4-e32d-11e6-9c6f-6927651257dd.png)

I think we can write our own conversion that is faster because it will be:

  - inlined by the compiler inside our module (the v8 function is exposed only in the v8 lib)
  - avoid extra calls to validate the object
  - will not create `v8::Isolates` since we know in this case we're in a single thread (the main event loop)