@springmeyer 2017-01-26
> Optimize v8::String -> const char * conversion

Profiling this showed that `malloc` was the main bottleneck. Because:

  - the [largest possible utf8 buffer can be predicted based on the string length](https://github.com/nodejs/node/blob/bfd3c7e626306cc5793618da2b56d37df338eb05/src/string_bytes.cc#L392), and
  - we can use the stack for most strings since they are small

This code is able to speed up the `node bench/compact-bench.js` from to `8000ms` to ~ `5000 ms`.

```diff
diff --git a/src/binding.cpp b/src/binding.cpp
index 6b83d2b..46b71b7 100644
--- a/src/binding.cpp
+++ b/src/binding.cpp
@@ -205,27 +205,58 @@ dawg_search_result compact_dawg_search(unsigned char* data, unsigned char* searc
     return output;
 }
 
+constexpr std::size_t arena_size = 1024;
+
 NAN_METHOD(CompactLookup) {
     // NOTE: these values are not validated since this
     // function is wrapped in JS
     v8::Local<v8::Object> bufferObj = info[0]->ToObject();
-    String::Utf8Value utf8_value(info[1].As<String>());
-
-    unsigned char* search = (unsigned char*) *utf8_value;
-    size_t search_length = utf8_value.length();
-
-    unsigned char* full_data = (unsigned char*) node::Buffer::Data(bufferObj);
-    unsigned char* data = full_data + DAWG_HEADER_SIZE;
-
-    dawg_search_result result = compact_dawg_search(data, search, search_length);
-
-    if (result.found) {
-        info.GetReturnValue().Set(result.final ? 2 : 1);
-        return;
-    } else {
-        info.GetReturnValue().Set(0);
-        return;
+    v8::Local<v8::Value> js_val = info[1];
+    std::uint32_t return_val = 1; 
+    // https://github.com/nodejs/node/commit/44a40325da4031f5a5470bec7b07fb8be5f9e99e
+    // https://github.com/nodejs/node/pull/1042
+    if (!js_val.IsEmpty()) {
+        v8::Local<v8::String> js_str = js_val->ToString();
+        if (!js_str.IsEmpty()) {
+            int js_str_len = js_str->Length();
+            if (js_str_len > 0) {
+                // Also passing v8::String::HINT_MANY_WRITES_EXPECTED flattens string
+                // but I've not enabled this yet as it does not clearly increase performance
+                const int flags =
+                    v8::String::NO_NULL_TERMINATION | v8::String::REPLACE_INVALID_UTF8;
+                unsigned char* full_data = (unsigned char*) node::Buffer::Data(bufferObj);
+                unsigned char* data = full_data + DAWG_HEADER_SIZE;
+                // max possible decoded utf length
+                // much faster than calling `str->Utf8Length();` to get exact length
+                // https://github.com/nodejs/node/blob/bfd3c7e626306cc5793618da2b56d37df338eb05/src/string_bytes.cc#L392
+                std::size_t len = (3 * js_str_len) + 1;
+                if (len > arena_size) {
+                    char * heap_string = static_cast<char *>(std::malloc(len));
+                    std::size_t utf8_length = js_str->WriteUtf8(heap_string, static_cast<int>(len), 0, flags);
+                    heap_string[utf8_length] = '\0';
+                    dawg_search_result result = compact_dawg_search(data, (unsigned char*)heap_string, utf8_length);
+                    if (result.found) {
+                        return_val = result.final ? 2 : 1;
+                    } else {
+                        return_val = 0;
+                    }
+                    free(heap_string);
+                } else {
+                    char arena[arena_size];
+                    std::size_t utf8_length = js_str->WriteUtf8(arena, static_cast<int>(len), 0, flags);
+                    arena[utf8_length] = '\0';
+                    dawg_search_result result = compact_dawg_search(data, (unsigned char*)arena, utf8_length);
+                    if (result.found) {
+                        return_val = result.final ? 2 : 1;
+                    } else {
+                        return_val = 0;
+                    }
+                }
+            }
+        }
     }
+    info.GetReturnValue().Set(return_val);
+    return;
 }
 
 class CompactIterator : public Nan::ObjectWrap {
```