@springmeyer 2017-01-27
I've implemented both optimizations per above in https://github.com/mapbox/dawg-cache/commit/b8b00552f69de6bf14e9a61c2ed84999e6693b29. With these changes then benchmark is now running in `4410ms`. I pushed this last commit to a new branch since it brought with it an API change to the iterator interface. @apendleton - did you say the es6 iterator interface is or is not currently used? Do I need to keep it working?