@springmeyer 2017-01-27
Just tested with carmen's `node bench/geocode.js` and I'm seeing this improvement:


#### master

```
# geocode
setup time 16713ms
geocode x 82.88 ops/sec ±5.75% (73 runs sampled) 
```

#### optimized dawg-cache

``` 
# geocode
setup time 16903ms
geocode x 92.25 ops/sec ±6.18% (70 runs sampled)
```

Needed this diff:

```diff
diff --git a/lib/util/dawg.js b/lib/util/dawg.js
index bf9e2e6..ec5d1c2 100644
--- a/lib/util/dawg.js
+++ b/lib/util/dawg.js
@@ -31,13 +31,11 @@ WriteCache.prototype.dump = function() {
 }
 
 var ReadCache = function(buf) {
-    _dawgCache.CompactDawg.call(this, buf);
+    this.cxx = new _dawgCache.CompactDawg(buf);
 }
 
-ReadCache.prototype = Object.create(_dawgCache.CompactDawg.prototype);
-
 ReadCache.prototype.hasPhrase = function(text, ender) {
-    return ender ? this.lookupPrefix(text) : this.lookup(text);
+    return ender ? this.cxx.lookupPrefix(text) : this.cxx.lookup(text);
 }
```